requirements.txt
flask
flask-cors
torch
transformers
sentence-transformers
faiss-cpu
numpy
huggingface_hub